feat: UI y funciones CRUD

+++ app/src/main/java/com/example/lab10/data/ProductoModel.kt
@@ -6,7 +6,7 @@ data class ProductoModel(
     @SerializedName("nombre")
     var nombre:String,
     @SerializedName("precio")
-    var precio:Int,
+    var precio:Float,
     @SerializedName("stock")
     var stock:Int,
     @SerializedName("pub_date")
